spark/spark-apps/requirements.txt
pyspark
kafka-python
psycopg2-binary
